Fixed undefined variable handling
Less substitutes variables in the toCSS phase of compilation, not the
parsing stage, by throwing an exception.  This exception was never
handled.

The result of this was that the error was reported as an ordinary
exception, with enough detail when node was being used, but when trireme
was being used, no detail was shown.

This fixes the problem by adding a try/catch around the conversion to
CSS, and handling errors from that in the same way as parsing errors are
handled.

Additionally, I did the following:

* Updated all scripted tests to use SBT 0.13.5 instead of 0.13.5-RC3
* Refactored error output test to run comprehensive assertions on error
  output, including line number, error message and file contents
* Made error output test run on both node and trireme, but only if node
  is installed
* Fixed an off by one error in loading the contents of files at the
  reported line number