Add dependency on nodeModules
In 460b3a5, I added the ability to import less sources from NPM modules.
However, the source tasks did not depend on the nodeModules task, so
builds could fail if the modules had not already been pulled into
$PROJECT_ROOT/node_modules.

This was hidden by the fact that most builds using NPM have other tasks
that fetch NPM modules, so the bug tended to show up only on the first
build in fresh checkouts.